Move normalize_chemical into the server package
normalize_chemical was being imported by the FastAPI chemicals router via
a sys.path hack into scripts/, but its runtime deps (pubchempy, rdkit,
requests) were never declared in server/pyproject.toml. The ImportError
was silently caught, making POST /api/chemicals/normalize return 503
under the new FastAPI layout.

Moved scripts/normalize_chemical.py to server/src/zapp_atlas/chem/normalize.py
so runtime code lives in the package that imports it. Sibling scripts and
the router now import from zapp_atlas.chem.normalize. Added pubchempy,
rdkit, and requests to server/pyproject.toml.
