+++ server/pyproject.toml
@@ -8,7 +8,10 @@ dependencies = [
     "boto3>=1.35",
     "fastapi>=0.135.1",
     "linkml",
+    "pubchempy>=1.0.4",
     "python-multipart>=0.0.9",
+    "rdkit>=2024.3",
+    "requests>=2.32",
     "sqlalchemy>=2.0",
     "uvicorn>=0.30.0",
 ]
